Assistant checkpoint: Add Buy/Rent toggle switch to property filters
Assistant generated file changes:
- src/components/properties/PropertyFilters.tsx: Add listing type toggle switch, Add toggle switch UI
- src/pages/Properties.tsx: Add listing type to initial filters

---

User prompt:

i would like from you to add something to the filters, a toogle switch. user should be able to toggle from Buy or Rent. those are the two options of the toggle. exactly like the one i showed you in the screenshoot. make sure you put it in top of the filters, like i wan it to be the very first one.